fixed64: codemod pass -- behavior-preserving Fixed->scalar casts tree-wide
Ran the clang-source-range codemod over all engine .cpp files: it wraps
each Fixed->scalar conversion the strict type now rejects in an explicit
cast to the exact target type (int/LONG/double/...), which is the same
truncation the non-strict build did implicitly -- behavior-identical in
both modes by construction, and any mis-wrap would be a compile error,
never a silent change. 540 casts across 59 files; the shrink guard
tripped on none.

Flag-off engine build verified green across all touched files; tests
green. This clears the mechanical bulk of the .cpp strict residue; what
remains per file is the non-mechanical set (network (LONG) args,
double<->fixed geometry, operator cases) fixed by hand as with po_man.

+++ src/zandronum/src/cl_commands.cpp
@@ -382,7 +382,7 @@ static void clientcommand_WriteMoveCommandToBuffer( CLIENT_MOVE_COMMAND_s moveCM
 		CLIENT_GetLocalBuffer( )->ByteStream.WriteShort( pCmd->ucmd.upmove );
 
 	CLIENT_GetLocalBuffer( )->ByteStream.WriteLong( players[consoleplayer].mo->angle );
-	CLIENT_GetLocalBuffer( )->ByteStream.WriteLong( players[consoleplayer].mo->pitch );
+	CLIENT_GetLocalBuffer( )->ByteStream.WriteLong( (int)(players[consoleplayer].mo->pitch) );
 	// [BB] Send the checksum of our ticcmd we calculated when we originally generated the ticcmd from the user input.
 	CLIENT_GetLocalBuffer( )->ByteStream.WriteLong( moveCMD.sdwChecksum );
 
@@ -825,8 +825,8 @@ void CLIENTCOMMANDS_InfoCheat( AActor* mobj, bool extended )
 void CLIENTCOMMANDS_WarpCheat( fixed_t x, fixed_t y )
 {
 	CLIENT_GetLocalBuffer( )->ByteStream.WriteByte( CLC_WARPCHEAT );
-	CLIENT_GetLocalBuffer( )->ByteStream.WriteLong( x );
-	CLIENT_GetLocalBuffer( )->ByteStream.WriteLong( y );
+	CLIENT_GetLocalBuffer( )->ByteStream.WriteLong( (int)(x) );
+	CLIENT_GetLocalBuffer( )->ByteStream.WriteLong( (int)(y) );
 }
 
 //*****************************************************************************
